WIP: Finishes testing first database methods.

+++ Awarean.BrayaOrtega.RinhaBackend.Q124/Models/Balance.cs
@@ -2,12 +2,12 @@ using System.Text.Json.Serialization;
 
 namespace Awarean.BrayaOrtega.RinhaBackend.Q124.Models;
 
-public readonly struct Balance(int total, int limite)
+public readonly struct Balance(long total, long limite)
 {
-    public int Total { get; } = total;
+    public long Total { get; } = total;
 
     [JsonPropertyName("data_extrato")]
     public DateTime DataExtrato { get; } = DateTime.Now;
 
-    public int Limite { get; } = limite;
+    public long Limite { get; } = limite;
 }
